requirements.txt
spacy>=3.7.0
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0
pytest>=7.0.0
tqdm>=4.66.0
ruff>=0.1.0
black>=23.0.0
mypy>=1.0.0
Flask>=3.0.0
Flask-CORS>=4.0.0
